Address architectural review: permissive graph semantics, GraphArtifact, u128 PlacementAddress
family-graph:
- Explicit 'coarse potential not certified' doc on build_graph and new
  test documenting why the graph is permissive, not a move witness
- Replace single aggregation test with four-property suite: one
  structural edge per pair, all types preserved, layout_weight=max,
  order independence of from_types
- Add GraphArtifact (project-owned adjacency list) so petgraph::Graph
  does not become the artifact format by inertia; includes roundtrip test

placement-codec:
- PlacementAddress u64 → u128: mixed-radix combinadics over promotion-
  heavy exact compositions (up to 9 queens per side) overflow 64 bits

+++ crates/placement-codec/src/lib.rs
@@ -11,7 +11,11 @@ use serde::{Deserialize, Serialize};
 use shakmaty::{Chess, Color, Square};
 
 /// Opaque combinadic rank identifying a non-pawn piece placement.
-pub type PlacementAddress = u64;
+///
+/// Mixed-radix combinadics over the free-square set. With up to 9 queens
+/// per side (after 8 promotions), the rank space can exceed 2^64 — u128
+/// is required. See DESIGN.md §Placement address.
+pub type PlacementAddress = u128;
 
 /// The context required to interpret a `PlacementAddress`.
 ///
